# perf(process): make sure we can take process spawning fast path on unix