--- a/Cargo.toml
+++ b/Cargo.toml
@@ -451,6 +451,8 @@ opt-level = 3
 opt-level = 3
 [profile.release.package.deno_web]
 opt-level = 3
+[profile.release.package.deno_process]
+opt-level = 3
 [profile.release.package.deno_websocket]
 opt-level = 3
 [profile.release.package.fastwebsockets]